Time: 263 ms (34.33%) | Memory: 0B (100.00%) - LeetSync

+++ 597-friend-requests-i-overall-acceptance-rate/friend-requests-i-overall-acceptance-rate.sql
@@ -3,6 +3,6 @@ SELECT
         WHEN (SELECT COUNT(*) FROM (SELECT DISTINCT sender_id, send_to_id FROM FriendRequest) AS req) = 0 THEN 0.00
         ELSE ROUND(
             (SELECT COUNT(*) FROM (SELECT DISTINCT requester_id, accepter_id FROM RequestAccepted) AS acc)::DECIMAL /
-            (SELECT COUNT(*) FROM (SELECT DISTINCT sender_id, send_to_id FROM FriendRequest) AS req), 2
+            (SELECT COUNT(*) FROM (SELECT DISTINCT sender_id, send_to_id FROM FriendRequest) AS req)::DECIMAL, 2
         )
     END AS accept_rate;
